// server/models/Experience.d.ts
import mongoose from "mongoose";
export declare const ExperienceModel: mongoose.Model<{
    id: number;
    title: string;
    description: string;
    price: string;
    location: string;
    imageUrl: string;
}, {}, {}, {}, mongoose.Document<unknown, {}, {
    id: number;
    title: string;
    description: string;
    price: string;
    location: string;
    imageUrl: string;
}, {}, mongoose.DefaultSchemaOptions> & {
    id: number;
    title: string;
    description: string;
    price: string;
    location: string;
    imageUrl: string;
} & {
    _id: mongoose.Types.ObjectId;
} & {
    __v: number;
}, mongoose.Schema<any, mongoose.Model<any, any, any, any, any, any>, {}, {}, {}, {}, mongoose.DefaultSchemaOptions, {
    id: number;
    title: string;
    description: string;
    price: string;
    location: string;
    imageUrl: string;
}, mongoose.Document<unknown, {}, mongoose.FlatRecord<{
    id: number;
    title: string;
    description: string;
    price: string;
    location: string;
    imageUrl: string;
}>, {}, mongoose.ResolveSchemaOptions<mongoose.DefaultSchemaOptions>> & mongoose.FlatRecord<{
    id: number;
    title: string;
    description: string;
    price: string;
    location: string;
    imageUrl: string;
}> & {
    _id: mongoose.Types.ObjectId;
} & {
    __v: number;
}>>;
//# sourceMappingURL=Experience.d.ts.map